refactor(ui): remove the dead pipeline UI now that discovery rides jobs (P7 S3.3)
After S3.2b moved NetworkDiscoveryCard onto the engine-scan job, nothing
consumed the legacy pipeline UI:
- delete usePipelineStatus.ts (765 lines: the hook + config-fetch + the
  port-intensity/timing-profile fetches that never had a UI consumer)
- delete PipelineProgress.tsx + PipelineProgress.stories.tsx (only the
  card used it)

The two still-live types (PipelineEvent / PipelineEventType) move into
useCardState.ts, their last consumer — the backend WebSocket/SSE event
router that parses incoming "pipeline" events. They (and that branch) go
away with the pipeline backend at retirement.

Net -~900 lines of dead frontend. tsc + biome clean, token-clean, 185
vitest green. No behavior change (deleted code had no runtime consumer).

Refs ADR-0007, SEED_S4_FOLD_PLAN.md
