--- ui/src/components/cards/PipelineProgress.tsx
@@ -1,241 +1,0 @@
-import { Check, Circle, FileText, Loader2, Scan, Search, Shield, X } from 'lucide-react';
-/**
- * Pipeline Progress Component
- *
- * Displays multi-phase discovery pipeline progress with:
- * - Phase steps with icons and status indicators
- * - Progress bar for current phase
- * - Current target being processed
- * - Elapsed time and cancel button
- */
-import type React from 'react';
-import { memo } from 'react';
-import { useTranslation } from 'react-i18next';
-import type { PipelineStatus } from '../../hooks/usePipelineStatus';
-import { button, cn, icon as iconTokens, radius, spacing } from '../../styles/theme';
-
-interface PipelineProgressProps {
-  status: PipelineStatus;
-  onCancel?: () => void;
-}
-
-// Phase metadata for display
-const PHASE_CONFIG: Record<
-  string,
-  { icon: React.ComponentType<{ className?: string }>; labelKey: string }
-> = {
-  enumeration: { icon: Search, labelKey: 'pipeline.phases.enumeration' },
-  resolution: { icon: FileText, labelKey: 'pipeline.phases.resolution' },
-  scanning: { icon: Scan, labelKey: 'pipeline.phases.scanning' },
-  assessment: { icon: Shield, labelKey: 'pipeline.phases.assessment' },
-};
-
-// Map phase name to display name (fallback for missing i18n)
-const PHASE_DISPLAY_NAMES: Record<string, string> = {
-  enumeration: 'Enumeration',
-  resolution: 'Resolution',
-  scanning: 'Scanning',
-  assessment: 'Assessment',
-};
-
-function formatDuration(ms: number): string {
-  // Fixes #957: Handle negative values from clock skew
-  const safeMs = Math.max(0, ms);
-  if (safeMs < 1000) {
-    return `${safeMs}ms`;
-  }
-  const seconds = Math.floor(safeMs / 1000);
-  if (seconds < 60) {
-    return `${seconds}s`;
-  }
-  const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = seconds % 60;
-  return `${minutes}m ${remainingSeconds}s`;
-}
-
-// Fixes #939: Simple hash for better key uniqueness instead of truncation
-function simpleHash(str: string): string {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    const char = str.charCodeAt(i);
-    hash = (hash << 5) - hash + char;
-    hash &= hash; // Convert to 32bit integer
-  }
-  return hash.toString(36);
-}
-
-/** Helper to render phase icon based on state */
-function renderPhaseIcon(
-  isComplete: boolean,
-  isCurrent: boolean,
-  isRunning: boolean,
-  Icon: React.ComponentType<{ className?: string }>,
-): React.ReactElement {
-  if (isComplete) {
-    return <Check className={iconTokens.size.sm} />;
-  }
-  if (isCurrent && isRunning) {
-    return <Loader2 className={cn(iconTokens.size.sm, 'animate-spin')} />;
-  }
-  return <Icon className={iconTokens.size.sm} />;
-}
-
-export const PipelineProgress: React.NamedExoticComponent<PipelineProgressProps> = memo(
-  function pipelineProgress({ status, onCancel }: PipelineProgressProps): React.ReactElement {
-    const { t } = useTranslation('cards');
-
-    const isRunning =
-      status.state === 'enumerating' ||
-      status.state === 'resolving' ||
-      status.state === 'scanning' ||
-      status.state === 'assessing';
-
-    // Fixes #913: Handle case where currentPhase is not in enabledPhases (returns -1)
-    // Default to 0 to show first phase as current instead of all phases as complete
-    const rawPhaseIndex = status.enabledPhases.indexOf(status.currentPhase);
-    const currentPhaseIndex = rawPhaseIndex >= 0 ? rawPhaseIndex : 0;
-
-    return (
-      <div className={cn('stack', spacing.pad.sm)}>
-        {/* Current phase header */}
-        <div className="flex-between">
-          <div className="flex items-center gap-compact">
-            {isRunning ? (
-              <Loader2 className={cn(iconTokens.size.sm, 'text-brand-primary animate-spin')} />
-            ) : null}
-            <span className="body-small font-medium text-text-primary">
-              {t('pipeline.phaseProgress', {
-                current: status.phaseNumber,
-                total: status.totalPhases,
-                defaultValue: `Phase ${status.phaseNumber} of ${status.totalPhases}`,
-              })}
-              : {PHASE_DISPLAY_NAMES[status.currentPhase] || status.currentPhase}
-            </span>
-          </div>
-          {isRunning && onCancel ? (
-            <button
-              type="button"
-              onClick={onCancel}
-              className={cn(
-                button.base,
-                button.size.sm,
-                button.variant.secondary,
-                'flex items-center gap-tight',
-              )}
-              aria-label={t('pipeline.cancel', { defaultValue: 'Cancel' })}
-            >
-              <X className={iconTokens.size.xs} />
-              <span className="hidden sm:inline">
-                {t('pipeline.cancel', { defaultValue: 'Cancel' })}
-              </span>
-            </button>
-          ) : null}
-        </div>
-        {/* Progress bar */}
-        <div className="stack-xs">
-          <div className={cn('h-2 bg-surface-sunken overflow-hidden', radius.default)}>
-            <div
-              className={cn('h-full bg-brand-primary transition-all duration-300', radius.default)}
-              style={{ width: `${Math.min(status.percentComplete, 100)}%` }}
-            />
-          </div>
-          <div className="flex justify-between caption text-text-muted">
-            <span>
-              {status.processedCount} / {status.totalCount}{' '}
-              {t('pipeline.devices', { defaultValue: 'devices' })}
-            </span>
-            <span>{Math.round(status.percentComplete)}%</span>
-          </div>
-        </div>
-        {/* Current target and timing */}
-        <div className="flex flex-wrap items-center gap-x-4 gap-y-1 caption text-text-muted">
-          {status.currentTarget ? (
-            <span>
-              {t('pipeline.scanning', { defaultValue: 'Scanning' })}:{' '}
-              <span className="font-mono text-text-secondary">{status.currentTarget}</span>
-            </span>
-          ) : null}
-          <span>
-            {t('pipeline.elapsed', { defaultValue: 'Elapsed' })}: {formatDuration(status.elapsedMs)}
-          </span>
-          {status.estimatedRemainMs > 0 && (
-            <span>
-              {t('pipeline.remaining', { defaultValue: 'Remaining' })}:{' '}
-              {formatDuration(status.estimatedRemainMs)}
-            </span>
-          )}
-        </div>
-        {/* Phase stepper */}
-        <div className="flex-between gap-tight pt-2 border-t border-surface-border">
-          {status.enabledPhases.map((phase, index) => {
-            const config = PHASE_CONFIG[phase];
-            const ICON = config?.icon || Circle;
-            const isComplete = index < currentPhaseIndex;
-            const isCurrent = index === currentPhaseIndex;
-            const isPending = index > currentPhaseIndex;
-
-            // Get duration if phase is complete
-            const duration = status.phaseDurations[phase];
-
-            return (
-              <div
-                key={phase}
-                className={cn(
-                  'flex flex-col items-center gap-tight flex-1',
-                  isPending && 'opacity-50',
-                )}
-              >
-                {/* Icon with status indicator */}
-                <div
-                  className={cn(
-                    'flex-center w-8 h-8',
-                    radius.full,
-                    isComplete && 'bg-status-success text-text-inverse',
-                    isCurrent && 'bg-brand-primary text-on-brand',
-                    isPending && 'bg-surface-sunken text-text-muted',
-                  )}
-                >
-                  {renderPhaseIcon(isComplete, isCurrent, isRunning, ICON)}
-                </div>
-                {/* Phase name */}
-                <span
-                  className={cn(
-                    'caption text-center',
-                    isCurrent ? 'text-text-primary font-medium' : 'text-text-muted',
-                  )}
-                >
-                  {PHASE_DISPLAY_NAMES[phase] || phase}
-                </span>
-                {/* Duration if complete */}
-                {duration !== undefined && (
-                  <span className="caption text-text-muted">{formatDuration(duration)}</span>
-                )}
-              </div>
-            );
-          })}
-        </div>
-        {/* Errors if any */}
-        {status.errors.length > 0 && (
-          <div
-            className={cn(
-              'pad-xs bg-status-error/10 border border-status-error/30',
-              radius.default,
-            )}
-          >
-            <span className="caption text-status-error font-medium">
-              {t('pipeline.errors', { defaultValue: 'Errors' })}:
-            </span>
-            <ul className="mt-tight space-y-0.5">
-              {/* Fixes #926, #939: Use error content hash for stable keys */}
-              {status.errors.map((error, i) => (
-                <li key={`${i}-${simpleHash(error)}`} className="caption text-status-error">
-                  {error}
-                </li>
-              ))}
-            </ul>
-          </div>
-        )}
-      </div>
-    );
-  },
-);
